requirements.txt
# Core dependencies
pyyaml>=6.0
netifaces>=0.11.0

# Network scanning (optional)
scapy>=2.4.5
python-nmap>=0.7.1

# GUI
tkinter  # Usually included with Python

# CLI enhancements
argparse  # Included with Python
colorama>=0.4.4
rich>=10.0.0

# For remote VM management
paramiko>=2.11.0